# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lethe v0.1 \u2014 internal codename",
    "description": "Closed-world localhost proof. Receipts are scoped signer assertions, not universal deletion proof.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "summary": "Accept Event",
        "operationId": "accept_event_api_v1_events_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Lethe-Event-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lethe-Event-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifecycleEvent"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/{version_id}": {
      "get": {
        "summary": "Get Graph",
        "operationId": "get_graph_api_v1_graph__version_id__get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/query": {
      "post": {
        "summary": "Query",
        "operationId": "query_api_v1_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/unsafe-query": {
      "post": {
        "summary": "Unsafe Query",
        "operationId": "unsafe_query_api_v1_demo_unsafe_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans": {
      "post": {
        "summary": "Scan",
        "operationId": "scan_api_v1_scans_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessments": {
      "post": {
        "summary": "Assess",
        "operationId": "assess_api_v1_assessments_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostureAssessmentReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/{run_id}": {
      "get": {
        "summary": "Get Receipt",
        "operationId": "get_receipt_api_v1_receipts__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/verify": {
      "post": {
        "summary": "Verify Receipts",
        "operationId": "verify_receipts_api_v1_receipts_verify_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventAccepted": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "workspace_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Workspace Id"
          },
          "environment_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Environment Id"
          },
          "event_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Event Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Run Id"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          },
          "gate_state": {
            "$ref": "#/components/schemas/LifecycleState"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "workspace_id",
          "environment_id",
          "event_id",
          "run_id",
          "gate_state"
        ],
        "title": "EventAccepted"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "delete",
          "correct",
          "expire",
          "permission_change"
        ],
        "title": "EventType"
      },
...
      "GraphResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "workspace_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Workspace Id"
          },
          "environment_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Environment Id"
          },
          "requested_version_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Requested Version Id"
          },
          "root_version_ids": {
            "items": {
              "type": "string",
              "maxLength": 512,
              "minLength": 1
            },
            "type": "array",
            "title": "Root Version Ids"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "workspace_id",
          "environment_id",
          "requested_version_id",
          "root_version_ids",
          "nodes",
          "edges"
        ],
        "title": "GraphResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LifecycleEvent": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "workspace_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Workspace Id"
          },
          "environment_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Environment Id"
          },
          "schema_version": {
            "type": "string",
            "const": "1",
            "title": "Schema Version",
            "default": "1"
          },
          "event_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Event Id"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "issuer_key_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Issuer Key Id"
          },
          "audience": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Audience"
          },
          "authority_ref": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Authority Ref"
          },
          "nonce": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Nonce"
          },
          "target_version_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Target Version Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "source_sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Source Sequence"
          },
          "policy_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Policy Version"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "effective_at": {
            "type": "string",
            "format": "date-time",
            "title": "Effective At"
          },
          "command_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Command Expires At"
          },
          "actor_ref": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Actor Ref"
          },
          "reason_code": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Reason Code"
          },
          "correction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CorrectionPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "permission_change": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PermissionChangePayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "workspace_id",
          "environment_id",
          "event_id",
          "idempotency_key",
          "issuer_key_id",
          "audience",
          "authority_ref",
          "nonce",
          "target_version_id",
          "event_type",
          "source_sequence",
          "policy_version",
          "occurred_at",
          "effective_at",
          "command_expires_at",
          "actor_ref",
          "reason_code"
        ],
        "title": "LifecycleEvent"
      },
      "LifecycleState": {
        "type": "string",
        "enum": [
          "active",
          "denied",
          "tombstoned"
        ],
        "title": "LifecycleState"
      },
...
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Query"
          },
          "purpose_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "supporting_version_ids": {
            "items": {
              "type": "string",
              "maxLength": 512,
              "minLength": 1
            },
            "type": "array",
            "title": "Supporting Version Ids"
          },
          "denied": {
            "type": "boolean",
            "title": "Denied"
          },
          "reason_codes": {
            "items": {
              "type": "string",
              "maxLength": 512,
              "minLength": 1
            },
            "type": "array",
            "title": "Reason Codes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "denied"
        ],
        "title": "QueryResponse"
      },
...
      "ReceiptVerifyRequest": {
        "properties": {
          "receipts": {
            "items": {
              "$ref": "#/components/schemas/ExecutionReceipt"
            },
            "type": "array",
            "title": "Receipts"
          },
          "expected_sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Sequence"
          },
          "expected_head": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 8,
            "title": "Expected Head"
          },
          "prior_trusted_head": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Prior Trusted Head"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "receipts",
          "expected_sequence",
          "expected_head"
        ],
        "title": "ReceiptVerifyRequest"
      },
      "ReceiptVerifyResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "verified_sequence": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Verified Sequence"
          },
          "computed_head": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed Head"
          },
          "reason_codes": {
            "items": {
              "type": "string",
              "maxLength": 512,
              "minLength": 1
            },
            "type": "array",
            "title": "Reason Codes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "valid",
          "verified_sequence"
        ],
        "title": "ReceiptVerifyResponse"
      },
...
      "RunOutcome": {
        "type": "string",
        "enum": [
          "succeeded",
          "partial",
          "failed",
          "unknown"
        ],
        "title": "RunOutcome"
      },
      "RunStatusResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "workspace_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Workspace Id"
          },
          "environment_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Environment Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Run Id"
          },
          "event_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Event Id"
          },
          "phase": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Phase"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer",
              "minimum": 0.0
            },
            "type": "object",
            "title": "Counts"
          },
          "failures": {
            "items": {
              "$ref": "#/components/schemas/ReceiptAction"
            },
            "type": "array",
            "title": "Failures"
          },
          "exclusions": {
            "items": {
              "$ref": "#/components/schemas/ReceiptExclusion"
            },
            "type": "array",
            "title": "Exclusions"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "workspace_id",
          "environment_id",
          "run_id",
          "event_id",
          "phase",
          "started_at"
        ],
        "title": "RunStatusResponse"
      },
...
      "ScanRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "workspace_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Workspace Id"
          },
          "environment_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Environment Id"
          },
          "root_version_ids": {
            "items": {
              "type": "string",
              "maxLength": 512,
              "minLength": 1
            },
            "type": "array",
            "title": "Root Version Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "workspace_id",
          "environment_id"
        ],
        "title": "ScanRequest"
      },
      "ScanResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "workspace_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Workspace Id"
          },
          "environment_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Environment Id"
          },
          "scan_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Scan Id"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ScanFinding"
            },
            "type": "array",
            "title": "Findings"
          },
          "denominator": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Denominator"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "workspace_id",
          "environment_id",
          "scan_id",
          "findings",
          "denominator",
          "completed_at"
        ],
        "title": "ScanResponse"
      },
...
    }
  }
}